+++ samples/DotNetCoreSample/Program.cs
@@ -84,13 +84,14 @@ namespace DotNetCoreSample
                     PKID=2,
                     SettingId = Guid.NewGuid(),
                     SettingName = "Setting2",
-                    SettingValue = "Value2"
+                    SettingValue = "Value2",
+                    Enabled = true
                 },
             };
             var csvFilePath = $@"{tempDirPath}\test.csv";
             entities.ToExcelFile(csvFilePath.Replace(".csv", ".xlsx"));
-            var entitiesT0 = ExcelHelper.ToEntityList<TestEntity>(csvFilePath.Replace(".csv", ".xlsx"));
             entities.ToCsvFile(csvFilePath);
+            var entitiesT0 = ExcelHelper.ToEntityList<TestEntity>(csvFilePath.Replace(".csv", ".xlsx"));
 
             var dataTable = entities.ToDataTable();
             dataTable.ToCsvFile(csvFilePath.Replace(".csv", ".datatable.csv"));
@@ -127,7 +128,7 @@ namespace DotNetCoreSample
                 .HasDescription("WeihanLi.Npoi test")
                 .HasSubject("WeihanLi.Npoi test");
 
-            setting.HasSheetConfiguration(0, "SystemSettingsList", 1);
+            setting.HasSheetConfiguration(0, "SystemSettingsList", 1, true);
 
             // setting.HasFilter(0, 1).HasFreezePane(0, 1, 2, 1);
 
@@ -155,13 +156,18 @@ namespace DotNetCoreSample
                 .HasColumnFormatter("yyyy-MM-dd HH:mm:ss");
 
             setting.Property(_ => _.CreatedBy)
+                .HasColumnInputFormatter(x => x += "_test")
                 .HasColumnIndex(4)
                 .HasColumnTitle("CreatedBy");
+
+            setting.Property(x => x.Enabled)
+                .HasColumnInputFormatter(val => "启用".Equals(val))
+                .HasColumnOutputFormatter(v => v ? "启用" : "禁用");
 
             setting.Property("HiddenProp")
                 .HasOutputFormatter((entity, val) => $"HiddenProp_{entity.PKID}");
 
-            // setting.Property(_ => _.PKID).Ignored();
+            setting.Property(_ => _.PKID).Ignored();
             setting.Property(_ => _.UpdatedBy).Ignored();
             setting.Property(_ => _.UpdatedTime).Ignored();
         }
@@ -188,6 +194,8 @@ namespace DotNetCoreSample
         public string UpdatedBy { get; set; }
 
         public DateTime UpdatedTime { get; set; }
+
+        public bool Enabled { get; set; }
     }
 
     [Sheet(SheetIndex = 0, SheetName = "TestSheet", AutoColumnWidthEnabled = true)]