+++ src/WeihanLi.Npoi/Abstract/IWorkbook.cs
@@ -1,0 +1,13 @@
+﻿namespace WeihanLi.Npoi.Abstract
+{
+    internal interface IWorkbook
+    {
+        int SheetCount { get; }
+
+        ISheet GetSheet(int sheetIndex);
+
+        ISheet CreateSheet(string sheetName);
+
+        byte[] ToBytes();
+    }
+}